src/requirements.txt
neo4j==5.15.0
openai
python-dotenv
flask
pyyaml